+++ CHANGELOG.md
@@ -2,16 +2,20 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 
 # [v1.5.0] (unreleased)
-### Notes
 
+### Notes
 - This release extends supports of the SDK to Oneview API1800.
 - Minor Bug Fixes
 
 ### Oneview Features supported
 - Enclosure
 - Enclosure Group
+- Ethernet Network
+- FCOE Network
+- FC Network
 - Hypervisor Cluster Profile
 - Hypervisor Manager
+- Scope
 - Server Certificate
 - Uplink Sets